Refactor to improve slot-type association and UI
The color is now based on the type name hash, no UI for it. The interactivity changed such that a selection is made first and then a type is associated by clicking its respective button. If no selection exists, the button opens the edit modal with rename and delete UI. The same modal is used for adding new types. Local storage is wrapped for easier manipulation of the data. The data is now saved as dates not slot numbers.