Remove no longer used variable
Since we don't need to look up the type object here, we just need to know its name which is already in the slot data itself, we don't need that array. The code that was there to look up the type object by name in that array was already removed.