Switch to annotating slots with free text as opposed to associating them with types
Initially the thing with having a predetermined set of types and marking slots with them seemed like a no brainer, but the more I use this, the more I realize what I really need is a way to annotate the slots and be able to see them at glance, without worrying about managing a set of types explicitly. The types fall out of the annotating when I do need them.